+++ Makefile
@@ -39,8 +39,9 @@ TEST_VECTOR_C_OBJS = $(TEST_VECTOR_C_SRCS:.c=.o)
 TEST_VECTOR_CXX_OBJS = $(TEST_VECTOR_CXX_SRCS:.cpp=.o)
 TEST_VECTOR_OBJS = $(TEST_VECTOR_C_OBJS) $(TEST_VECTOR_CXX_OBJS)
 
-LEXER_HEADER = $(SRC_DIR)/lex.yy.h $(SRC_DIR)/token.h $(SRC_DIR)/yacc_mock.h $(SRC_DIR)/lexer_adapter.h
-LEXER_OBJS = $(SRC_DIR)/lex.yy.o $(SRC_DIR)/lexer_adapter.o $(SRC_DIR)/yacc_mock.o
+LEXER_DIR = $(SRC_DIR)/lexer
+LEXER_HEADER = $(LEXER_DIR)/lex.yy.h $(LEXER_DIR)/token.h $(LEXER_DIR)/yacc_mock.h $(LEXER_DIR)/lexer_adapter.h
+LEXER_OBJS = $(LEXER_DIR)/lex.yy.o $(LEXER_DIR)/lexer_adapter.o $(LEXER_DIR)/yacc_mock.o
 TEST_LEXER_DIR = $(TESTS_DIR)/lexer_test
 TEST_LEXER_C_SRCS = $(wildcard $(TEST_LEXER_DIR)/*.c)
 TEST_LEXER_CXX_SRCS = $(wildcard $(TEST_LEXER_DIR)/*.cpp)
@@ -97,8 +98,8 @@ $(GTEST_LIB): $(GTEST_OBJS)
 $(GTEST_DIR)/%.o: $(GTEST_DIR)/%.cc
 	$(CXX) $(CXXFLAGS) $(GTEST_INCLUDE) $(GTEST_FLAGS) -c $< -o $@
 
-$(SRC_DIR)/lex.yy.c $(SRC_DIR)/lex.yy.h: $(SRC_DIR)/lexer.l
-	$(LEX) --header-file=$(SRC_DIR)/lex.yy.h -o $(SRC_DIR)/lex.yy.c $(SRC_DIR)/lexer.l
+$(LEXER_DIR)/lex.yy.c $(LEXER_DIR)/lex.yy.h: $(LEXER_DIR)/lexer.l
+	$(LEX) --header-file=$(LEXER_DIR)/lex.yy.h -o $(LEXER_DIR)/lex.yy.c $(LEXER_DIR)/lexer.l
 
 $(TEST_VECTOR_DIR)/vector_int_test.out: $(VECTOR_OBJ) $(TEST_VECTOR_OBJS) $(GTEST_LIB)
 	$(CXX) $(CXXFLAGS) $^ -o $@ $(GTEST_LDFLAGS)
@@ -126,6 +127,9 @@ $(SRC_DIR)/%.o: $(SRC_DIR)/%.c
 
 $(SRC_DIR)/%.o: $(SRC_DIR)/%.cpp
 	$(CXX) $(CXXFLAGS) $(LLVM_CXXFLAGS) -c $< -o $@
+
+$(LEXER_DIR)/%.o: $(LEXER_DIR)/%.c $(LEXER_HEADER)
+	$(CC) $(CFLAGS) $(LLVM_CFLAGS) -I$(SRC_DIR) -c $< -o $@
 
 clean:
 	$(RM) $(TESTS_OBJS) $(UNITTESTS) $(GTEST_OBJS) $(VECTOR_OBJ) $(TEST_VECTOR_OBJS) $(KMC89_OBJS)
