+++ src/lexer/lexer_adapter.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "lex.yy.h"
-#include "yacc_mock.h"
-#include "lexer_adapter.h"
+#include "lexer/lex.yy.h"
+#include "lexer/yacc_mock.h"
+#include "lexer/lexer_adapter.h"
 
 void yyin_from_file(const char *filename) {
   FILE* fp = fopen(filename, "r");
