+++ src/lexer/yacc_mock.c
@@ -1,0 +1,3 @@
+#include "lexer/yacc_mock.h"
+
+char *yylval;
