--- src/yacc_mock.c
@@ -1,3 +1,0 @@
-#include "yacc_mock.h"
-
-char *yylval;
